Add possibility to use multiple user snippet files
IMPORTANT: It is backwards-compatible. I first check if the variable
is a string (used until this PR) and keep the existing code for that
(except adding the runtime command for loading the snippet file). Only
if the variable is a list a for loop for loading multiple
snippet files is used.